Fixing fitness function

+++ BinPackingProblem/src/com/packingproblem/entity/Bag.kt
@@ -1,10 +1,10 @@
 package com.packingproblem.main
 
 class Bag (val items: IntArray, var fitness: Double, var totalWeight: Int, var totalValue: Int,
-		   var factible : Boolean, var percentOfFitness: Long, var totalItems : Int) {
+		   var factible : Boolean, var percentOfFitness: Double, var totalItems : Int) {
 	
 	constructor(items: IntArray, totalWeight: Int, totalValue: Int, totalItems: Int) :
-			this(items, 0.0, totalWeight, totalValue, totalWeight <= 120, 0, totalItems)
+			this(items, 0.0, totalWeight, totalValue, totalWeight <= 120, 0.0, totalItems)
 	
 	override fun equals(other: Any?) : Boolean {
 		if (this === other) return true
